% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFnest.R
\name{myFnest}
\alias{myFnest}
\title{myFnest}
\usage{
myFnest(objc, objr, alpha = 0.05)
}
\arguments{
\item{objc}{an `lm()` object for the complete model}

\item{objr}{an `lm()` object for the reduced model}

\item{alpha}{the alpha level}
}
\value{
a list containing the nested F-statistic, the F_alpha value, and result of hyp test
}
\description{
A function which tests the comparison of nested models (Hierarchical F-Test)
}
\examples{
\dontrun{myFnest(ylm, zlm, alpha=0.05)}
}
