% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLAreg.R
\name{myLAreg}
\alias{myLAreg}
\title{myLAreg}
\usage{
myLAreg(obj)
}
\arguments{
\item{obj}{an `lm()` object}
}
\value{
a named list with the following: model summary, n (num observations), k (num beta parameters), beta estimates, RSS, s-squared, s, and the Design & Response Matrices.
}
\description{
A function which takes an `lm()` object and produces a list of fun and interesting values which Regression covets
}
\examples{
\dontrun{myLAreg(ylm)}
}
