% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPartialF.R
\name{myPartialF}
\alias{myPartialF}
\title{myPartialF}
\usage{
myPartialF(objr, objc, alpha = 0.05)
}
\arguments{
\item{objr}{an `lm()` object of the reduced model}

\item{objc}{an `lm()` object of the complete model}

\item{alpha}{alpha level, default is 0.05}
}
\value{
a list containing the ANOVA summary, F_Statistic, p-value, alpha level, and test conclusion.
}
\description{
A function which runs a Partial F-Test using the `anova()` function.
}
\examples{
\dontrun{myPartialF(lm1, lm2, alpha=0.05)}
}
